Adding exercises to weekdays working

+++ app/src/main/java/com/example/assignemnt_fit/model/ExerciseWeekDayJoin.kt
@@ -11,6 +11,6 @@ import androidx.room.ForeignKey
     ]
 )
 data class ExerciseWeekDayJoin(
-    val exerciseId: Int,
-    val weekDayId: Int
+    val exerciseId: Long,
+    val weekDayId: Long
 )
